NetworkSettings: Don't error out with TD-SCDMA values
This adds persistence and the ability to read TD-SCDMA values if
they are set. Both Network configuration and the QS tile toggle
now cycle these modes correctly
Devices supporting TD-SCDMA need to overlay the config_enabled_tdscdma
boolean configuration value to "true" for the proper network modes
to be presented to the user.

The ro.telephony.default_network should also be changed to a TD-SCDMA
value (20 or 18, depending on whether LTE is also supported)

Change-Id: I3a88a01aa9c5c111101c53388d5ba76a99cd0df6