Call recording encoder/format choice 2/3
PS1: Adds a user option between WB AMR (the default) or HE-AAC for the
 encoding of call recording audio. Also changes the output format of the
 file appropriately (WB AMR now uses the WB_AMR output format and HE_AAC
 chooses the MPEG_4 output. File extensions are adjust accordingly as well).

This was in reponse to people complaining the quality of call recording
 audio is poor. It's noticibly better using HE-AAC, but filesizes are larger.

PS2: Made recommended changes to the method of determining whether or not
 call recording is enabled and fixed an erroneous array name.

PS3: Fixed errors in the PS2, should show proper changes now.

PS4: Removed whitespace

PS5: Whitespace again

PS6: Fixed getting the identifier, now works as intended. Altered the
 default value that the preference setting gets in order to show the
 default, untouched, preference summary

Change-Id: I31f4a57042cf9e221ea2292922f071c955170e1e